Add MaskedLM for keras_nlp
PiperOrigin-RevId: 332506544
